P2: 5.Placing baby image
Placing baby image on the X and Y axis. Using jquery animation to move the baby on the X axis from the left to right but keeping the left as an initial area of position. Also using jquery animation and to move the baby on the Y axis from the top to bottom and keeping the top to be the initial position.